\name{MissingParafac-package}
\alias{MissingParafac-package}
\alias{MissingParafac}
\docType{package}
\title{
\packageTitle{MissingParafac}
}
\description{
\packageDescription{MissingParafac}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MissingParafac}
\packageIndices{MissingParafac}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MissingParafac}

Maintainer: \packageMaintainer{MissingParafac}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
